feat: implementando métodos de extensão
